backend-temp/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
boto3==1.39.7
botocore==1.39.7
certifi==2025.7.9
charset-normalizer==3.4.2
click==8.2.1
distro==1.9.0
Flask==3.1.1
flask-cors==6.0.1
gunicorn
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
MarkupSafe==3.0.2
numpy==2.3.1
openai==1.97.0
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
requests==2.32.4
s3transfer==0.13.0
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
pytz==2025.1
user-agents
